修复演示BUG

+++ wind-server/wind-common/wind-common-base/src/main/java/com/sunseagear/common/mvc/service/impl/TreeCommonServiceImpl.java
@@ -76,6 +76,9 @@ public class TreeCommonServiceImpl<M extends BaseTreeMapper<T>, T extends Serial
 
     @Override
     public boolean deleteById(Serializable id) {
+        if (isDemo){
+            return true;
+        }
         T entity = selectById(id);
         baseMapper.deleteSunTree(entity.makeSelfAsNewParentIds());
         return super.deleteById(id);
